+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+# domino x.x.x (not yet released)
+
 # domino 2.1.2 (14 Feb 2019)
 * Allow writable Element constructors unless __domino_frozen__ is set to true (#138)
 * Bug fix for CSS `$=` selector. (#135)