+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# domino x.x.x (not yet released)
+# domino 2.1.3 (6 Mar 2019)
 * Bug fix for CSS `$=` selector and for matches on root `<html>` element.
 * Renamed CSS `:matches` to `:is`
   ( https://github.com/w3c/csswg-drafts/issues/3258 )